+++ iphone/Classes/TiUITableViewProxy.m
@@ -133,12 +133,12 @@ USE_VIEW_FOR_CONTENT_HEIGHT
 -(TiUITableViewRowProxy*)rowForIndex:(NSInteger)index section:(NSInteger*)section
 {
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	int sectionIdx = 0;
 	
 	for (TiUITableViewSectionProxy *sectionProxy in sections)
 	{
-		int rowCount = [sectionProxy rowCount];
+		NSInteger rowCount = [sectionProxy rowCount];
 		if (rowCount + current > index)
 		{
 			if (section!=nil)
@@ -163,11 +163,11 @@ USE_VIEW_FOR_CONTENT_HEIGHT
 	}
 	int section = 0;
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	
 	for (TiUITableViewSectionProxy * thisSection in sections)
 	{
-		int rowCount = [thisSection rowCount];
+		NSInteger rowCount = [thisSection rowCount];
 		if (rowCount + current > index)
 		{
 			return [NSIndexPath indexPathForRow:row inSection:section];
@@ -263,7 +263,7 @@ USE_VIEW_FOR_CONTENT_HEIGHT
 -(TiUITableViewSectionProxy*)sectionForIndex:(NSInteger)index row:(TiUITableViewRowProxy**)rowOut
 {
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	int sectionIdx = 0;
 	
 	TiUITableViewRowProxy *rowProxy = nil;
@@ -271,7 +271,7 @@ USE_VIEW_FOR_CONTENT_HEIGHT
 	
 	for (sectionProxy in sections)
 	{
-		int rowCount = [sectionProxy rowCount];
+		NSInteger rowCount = [sectionProxy rowCount];
 		if (rowCount + current > index)
 		{
 			rowProxy = [sectionProxy rowAtIndex:row];
@@ -392,7 +392,7 @@ USE_VIEW_FOR_CONTENT_HEIGHT
         
         for (sectionProxy in sections)
         {
-            int rowCount = [sectionProxy rowCount];
+            NSInteger rowCount = [sectionProxy rowCount];
             if (rowCount + current > index)
             {
                 rowProxy = [sectionProxy rowAtIndex:row];
@@ -531,7 +531,7 @@ USE_VIEW_FOR_CONTENT_HEIGHT
         TiUITableViewSectionProxy *newSection = [self sectionWithHeader:header table:table];
         
         // Insert the new section into the array - but, exactly WHERE we insert depends.
-        int sectionIndex = [sections indexOfObject:section];
+        NSInteger sectionIndex = [sections indexOfObject:section];
         if (row.row != 0) {
             sectionIndex++;
         }
@@ -541,7 +541,7 @@ USE_VIEW_FOR_CONTENT_HEIGHT
 		
         // Thanks to how we track sections, we also need to manually update the index
         // of each section in the array after where the insert will be.
-        for (int i=sectionIndex; i < [sections count]; i++) {
+        for (NSUInteger i=sectionIndex; i < [sections count]; i++) {
             TiUITableViewSectionProxy *updateSection = [sections objectAtIndex:i];
             updateSection.section = updateSection.section + 1;
         }
@@ -614,11 +614,11 @@ USE_VIEW_FOR_CONTENT_HEIGHT
         newSection.section = section.section + 1;
         
         // Insert the new section into the array
-        int sectionIndex = [sections indexOfObject:section] + 1;
+        NSUInteger sectionIndex = [sections indexOfObject:section] + 1;
         
         // Thanks to how we track sections, we also need to manually update the index
         // of each section in the array after where the insert will be.
-        for (int i=sectionIndex; i < [sections count]; i++) {
+        for (NSUInteger i=sectionIndex; i < [sections count]; i++) {
             TiUITableViewSectionProxy *updateSection = [sections objectAtIndex:i];
             updateSection.section = updateSection.section + 1;
         }
@@ -887,7 +887,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
     [(TiUITableView *)[self view] refreshSearchControllerUsingReload:YES];
 }
 
--(int)sectionCount
+-(NSUInteger)sectionCount
 { //TODO: Shouldn't this be in the main thread, too?
 	return [sections count];
 }
@@ -912,7 +912,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 -(void)appendSection:(id)args
 {
 	//Step one: sanity
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	if(argCount < 1){
 		[self throwException:TiExceptionNotEnoughArguments
 				   subreason:@"expected 1 argument, received none" location:CODELOCATION];
@@ -924,7 +924,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 	NSMutableArray * sectionArray = nil;
 
 	if ([appendum isKindOfClass:[NSArray class]]) {
-		int arrayCount = [appendum count];
+		NSUInteger arrayCount = [appendum count];
 		if (arrayCount < 1) {
 			[self throwException:TiExceptionNotEnoughArguments subreason:@"array must not be empty" location:CODELOCATION];
 		}
@@ -956,7 +956,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 		NSRange sectionRange = NSMakeRange([sections count], 1);
 		if (sectionArray != nil){
 			sectionRange.length = [sectionArray count];
-			int sectionIndex = sectionRange.location;
+			NSUInteger sectionIndex = sectionRange.location;
 			for (TiUITableViewSectionProxy * thisSection in sectionArray) {
 				[thisSection setSection:sectionIndex++];
 			}
@@ -991,7 +991,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 -(void)deleteSection:(id)args
 {
 	//Step one: sanity
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	if(argCount < 1){
 		[self throwException:TiExceptionNotEnoughArguments
 				   subreason:@"expected 1 argument, received none" location:CODELOCATION];
@@ -1030,8 +1030,8 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 {
 	[self rememberSection:section];
 	TiThreadPerformOnMainThread(^{
-		int oldSectionCount = [sections count];
-		int boundSectionIndex = MIN(sectionIndex, oldSectionCount);
+		NSUInteger oldSectionCount = [sections count];
+		NSUInteger boundSectionIndex = MIN(sectionIndex, oldSectionCount);
 		[section setSection:boundSectionIndex];
 		[sections insertObject:section atIndex:boundSectionIndex];
 
@@ -1052,7 +1052,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 
 -(void)insertSectionAfter:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 
 	NSDictionary * options = nil;
@@ -1078,7 +1078,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 
 -(void)insertSectionBefore:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 	
 	NSDictionary * options = nil;
@@ -1104,7 +1104,7 @@ DEFINE_DEF_PROP(scrollsToTop,[NSNumber numberWithBool:YES]);
 
 -(void)updateSection:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 	
 	NSDictionary * options = nil;
